Change to using 'unsigned long int' for the allocation sizes of memory so we can support large pointers on x86_64 and aix.
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@4324 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/general/debug.h
@@ -178,7 +178,7 @@ is swallowed with the call USE_PARAMETER(dummy_void); at the start of function.
 #endif /* defined (USE_PARAMETER_ON) */
 
 #if !defined (OPTIMISED)
-char *allocate(unsigned size,char *file_name,int line_number,char *type);
+char *allocate(unsigned long int size,char *file_name,int line_number,char *type);
 /*******************************************************************************
 LAST MODIFIED : 7 January 1998
 
@@ -194,7 +194,7 @@ DESCRIPTION :
 Wrapper for free.
 ==============================================================================*/
 
-char *reallocate(char *ptr,unsigned size,char *file_name,int line_number,
+char *reallocate(char *ptr,unsigned long int size,char *file_name,int line_number,
 	char *type);
 /*******************************************************************************
 LAST MODIFIED : 7 January 1998
